[#25081] Extract attachments into a concern
