+++ app/models/stored_file.rb
@@ -1,4 +1,6 @@
 class StoredFile < ActiveRecord::Base
+  include DownloadableFile
+
   belongs_to              :product
   belongs_to              :creator, :class_name => 'User', :foreign_key => 'created_by'
   belongs_to              :order_detail
@@ -6,11 +8,6 @@ class StoredFile < ActiveRecord::Base
   validates_presence_of   :product_id,      :if => lambda {|o| o.file_type == 'info' || o.file_type == 'template'}
   validates_presence_of   :order_detail_id, :if => lambda {|o| o.file_type == 'template_result' || o.file_type == 'sample_result'}
   validates_inclusion_of  :file_type, :in => %w(info template template_result sample_result import_error import_upload)
-  has_attached_file       :file, Settings.paperclip.to_hash
-  validates_attachment_presence :file
-
-  # TODO Limit attachment types for safe uploads
-  do_not_validate_attachment_file_type :file
 
   scope :info,              :conditions => {:file_type => 'info'}
   scope :template,          :conditions => {:file_type => 'template'}
@@ -19,13 +16,7 @@ class StoredFile < ActiveRecord::Base
   scope :import_upload,     :conditions => {:file_type => 'import_upload'}
   scope :import_error,      :conditions => {:file_type => 'import_error'}
 
-  def download_url
-    file.send(:directory).files.get_url(
-      file.path,
-      10.seconds.from_now,
-      query: { "response-content-disposition" => "attachment" },
-    )
-  end
+  validates_attachment_presence :file
 
   def read
     Paperclip.io_adapters.for(file).read
