+++ config/application.rb
@@ -22,6 +22,7 @@ module Nucore
     config.autoload_paths += Dir["#{config.root}/lib/**/"]
     config.autoload_paths += Dir["#{config.root}/app/models/**/"]
     config.autoload_paths += Dir["#{config.root}/app/controllers/concerns"]
+    config.autoload_paths += Dir["#{config.root}/app/models/concerns"]
     config.autoload_paths += Dir["#{config.root}/app/validators"]
 
     # Only load the plugins named here, in the order given (default is alphabetical).
